# chatgpt_openapi_schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenFinder - Universal AI LinkedIn Scout & Recruiter Intelligence API",
    "description": "High-precision LinkedIn hiring post finder, candidate resume profile extractor, and recruiter outreach suite for ChatGPT Custom GPTs. Discovers verified /posts/ URLs with exact minute-level freshness and ATS candidate fit scoring.",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "https://openfinder.onrender.com"
    }
  ],
  "paths": {
    "/api/upload-resume": {
      "post": {
        "summary": "Upload and Parse Candidate PDF Resume",
        "description": "Uploads candidate PDF resume, extracts structured skills, experience level, and target roles, stores a persistent candidate profile in SQLite, and returns a unique candidate_profile_id.",
        "operationId": "uploadResume",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Candidate PDF resume file"
                  }
                },
                "required": ["file"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Candidate profile successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string" },
                    "candidate_profile_id": { "type": "string" },
                    "candidate_name": { "type": "string" },
                    "seniority_level": { "type": "string" },
                    "years_of_experience": { "type": "integer" },
                    "primary_role": { "type": "string" },
                    "top_skills": { "type": "array", "items": { "type": "string" } },
                    "target_roles": { "type": "array", "items": { "type": "string" } }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/candidate-profile/{profile_id}": {
      "get": {
        "summary": "Retrieve Stored Candidate Profile",
        "description": "Retrieves structured candidate skills, experience, and target roles using a candidate_profile_id.",
        "operationId": "getCandidateProfile",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Unique candidate profile ID (e.g. 'prof_a1b2c3d4e5f6')"
          }
        ],
        "responses": {
          "200": {
            "description": "Candidate profile details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string" },
                    "candidate_profile": { "type": "object" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/search-opportunities": {
      "get": {
        "summary": "Search Verified LinkedIn Recruiter Hiring Opportunities",
        "description": "Canonical search operation. Discovers verified genuine LinkedIn hiring /posts/ with exact freshness filters, directional hiring intent, and opportunity ranking against candidate profile.",
        "operationId": "searchOpportunities",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "default": "React Developer"
            },
            "description": "Job role or technical skill keywords (e.g. 'React Developer', 'MERN Stack', 'Python Backend')"
          },
          {
            "name": "location",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "India"
            },
            "description": "City or Region (e.g. 'Bangalore', 'Chennai', 'Remote', 'India')"
          },
          {
            "name": "timeframe",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "past-24h",
              "enum": ["past-1h", "past-4h", "past-12h", "past-24h", "past-7d"]
            },
            "description": "Exact freshness window: 'past-1h' (last 1 hr), 'past-4h' (4 hrs), 'past-12h' (12 hrs), 'past-24h' (24 hrs), or 'past-7d' (7 days)"
          },
          {
            "name": "max_results",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Max opportunities to return (1-30)"
          },
          {
            "name": "remote_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "Filter for remote roles only"
          },
          {
            "name": "candidate_profile_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Optional stored candidate profile ID for ATS skill and experience fit ranking"
          }
        ],
        "responses": {
          "200": {
            "description": "List of ranked verified hiring posts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string" },
                    "query": { "type": "string" },
                    "location": { "type": "string" },
                    "timeframe": { "type": "string" },
                    "count": { "type": "integer" },
                    "candidate_profile_id": { "type": "string" },
                    "results": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "company": { "type": "string" },
                          "role": { "type": "string" },
                          "author": { "type": "string" },
                          "author_type": { "type": "string" },
                          "location": { "type": "string" },
                          "posted_time": { "type": "string" },
                          "age_minutes": { "type": "integer" },
                          "hiring_intent": { "type": "string" },
                          "post_quality_score": { "type": "integer" },
                          "candidate_match_score": { "type": "integer", "nullable": true },
                          "final_rank_score": { "type": "integer" },
                          "ranking_summary": { "type": "string" },
                          "ranking_reasons": { "type": "array", "items": { "type": "string" } },
                          "recruiter_emails": { "type": "array", "items": { "type": "string" } },
                          "contact_numbers": { "type": "array", "items": { "type": "string" } },
                          "skills": { "type": "array", "items": { "type": "string" } },
                          "post_url": { "type": "string" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/generate-pitch": {
      "post": {
        "summary": "Generate Customized Recruiter Outreach Pitch Suite",
        "description": "Generates 4 personalized recruiter outreach templates (LinkedIn Connection Note <300 chars, InMail, Formal Cover Email, Day-3 Follow-up).",
        "operationId": "generatePitch",
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "job_title": {
                    "type": "string",
                    "description": "Target job title (e.g. 'React Developer')"
                  },
                  "company_name": {
                    "type": "string",
                    "default": "Hiring Team",
                    "description": "Target hiring company"
                  },
                  "matched_skills": {
                    "type": "string",
                    "default": "React, TypeScript",
                    "description": "Comma-separated matched technical skills"
                  },
                  "candidate_name": {
                    "type": "string",
                    "default": "Candidate",
                    "description": "Candidate name for sign-off"
                  },
                  "candidate_exp_years": {
                    "type": "integer",
                    "default": 2,
                    "description": "Candidate total years of experience"
                  }
                },
                "required": ["job_title"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Outreach pitch suite generated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string" },
                    "job_title": { "type": "string" },
                    "company": { "type": "string" },
                    "pitches": { "type": "object" }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}